encode node and topic names always as UTF8
This is to prevent errors caused by null bytes
in C strings depending on the Windows codepage.
Furthermore, it prevents nodes on different platforms
from failing to find each others topics since they where
encoded differently.
The encoding cant be done with UnmanagedType.LPUTF8Str
since it is not available in Netstandard 2.0